backend/requirements.txt
Django>=4.2
djangorestframework
stripe
python-dotenv
gunicorn